Add the inner class UpdaterThread
This implements a single run() member function, which repeatedly
posts a given function (as a Runnable) to the (Handler of the)
given Looper, every so many milli seconds.

This was implemented as a class derived from Thread, rather than
as a Runnable that woudl eventually be wrapped in a Thread constructor,
because I wanted to have some visibility of the Thread, to be able to
stop it co-operatively.