Set the module "lib" to depend on "xapi2" but to do so in the "api" mode.
This means that the published interface for "xapi2" is used by "lib" but
also that it is propagated upwards from "lib" itself, along with the
published interface for the latter.

This is exactly what we want.

In particular, the host program cannot (and, by design, must not) see any
api other than that published by lib. If a host program requires that api
published by xapi2 (and later, xapi1), it will do so by using that api
published by lib (not least because those two xapiN modules will have lost
their individual identities when being compiled into the apk that will
eventually be the lib distributable)